requirements.txt
greenlet==2.0.2
numpy==1.24.3
pandas==2.0.2
psycopg2==2.9.7
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.6.2
tzdata==2023.3
